+++ include/viarulegenerate.h
@@ -8,22 +8,22 @@ namespace phydb {
 class ViaRuleGenerateLayer {
   public:
     string layer_name_;
-    Rect2D<float> rect_;
-    Size2D<float> spacing_;
-    Size2D<float> enclosure_;
+    Rect2D<double> rect_;
+    Size2D<double> spacing_;
+    Size2D<double> enclosure_;
 
     ViaRuleGenerateLayer() {}
     ViaRuleGenerateLayer(string &layerName) : layer_name_(layerName) {}
     
     void SetLayerName(const string& layer_name);
-    void SetRect(float lx, float ly, float ux, float uy);
-    void SetSpacing(float x, float y);
-    void SetEnclosure(float x, float y);
+    void SetRect(double lx, double ly, double ux, double uy);
+    void SetSpacing(double x, double y);
+    void SetEnclosure(double x, double y);
 
     string GetLayerName() const;
-    Rect2D<float> GetRect() const;
-    Size2D<float> GetSpacing() const;
-    Size2D<float> GetEnclosure() const;
+    Rect2D<double> GetRect() const;
+    Size2D<double> GetSpacing() const;
+    Size2D<double> GetEnclosure() const;
 };
 
 class ViaRuleGenerate {
