+++ src/lefvia.cpp
@@ -10,9 +10,9 @@ void LefVia::UnsetDefault() {
     is_default_ = false;
 }
  
-void LefVia::SetLayerRect(const string& layer_name0, const vector<Rect2D<float>>& rects0,
-        const string& layer_name1, const vector<Rect2D<float>>& rects1,
-        const string& layer_name2, const vector<Rect2D<float>>& rects2) {
+void LefVia::SetLayerRect(const string& layer_name0, const vector<Rect2D<double>>& rects0,
+        const string& layer_name1, const vector<Rect2D<double>>& rects1,
+        const string& layer_name2, const vector<Rect2D<double>>& rects2) {
     layer_rects_[0] = LayerRect(layer_name0, rects0);
     layer_rects_[1] = LayerRect(layer_name1, rects1);
     layer_rects_[2] = LayerRect(layer_name2, rects2);
