+++ src/spacingtable.cpp
@@ -27,59 +27,59 @@ int SpacingTable::GetNRow( ) const {
     return n_row_;
 }
 
-void SpacingTable::SetParallelRunLengthVec(const vector<float>& v) {
+void SpacingTable::SetParallelRunLengthVec(const vector<double>& v) {
     assert(v.size() == n_col_);
     parallel_run_length_ = v;
 }
 
-void SpacingTable::SetWidthVec(const vector<float>& v) {
+void SpacingTable::SetWidthVec(const vector<double>& v) {
     assert(v.size() == n_row_);
     width_ = v;
 }
 
-void SpacingTable::SetSpacingVec(const vector<float>& v) {
+void SpacingTable::SetSpacingVec(const vector<double>& v) {
     assert(v.size() == n_row_ * n_col_);
     spacing_ = v;
 }
 
-vector<float>& SpacingTable::GetParallelRunLengthVec() {
+vector<double>& SpacingTable::GetParallelRunLengthVec() {
     return parallel_run_length_;
 }
 
-vector<float>& SpacingTable::GetWidthVec() {
+vector<double>& SpacingTable::GetWidthVec() {
     return width_;
 }
 
-vector<float>& SpacingTable::GetSpacingVec() {
+vector<double>& SpacingTable::GetSpacingVec() {
     return spacing_;
 }
 
-void SpacingTable::SetParallelRunLengthAt(int col, float val) {
+void SpacingTable::SetParallelRunLengthAt(int col, double val) {
     assert(col < n_col_);
     parallel_run_length_[col] = val;
 }
 
-void SpacingTable::SetWidthAt(int row, float val) {
+void SpacingTable::SetWidthAt(int row, double val) {
     assert(row < n_row_);
     width_[row] = val;
 }
 
-void SpacingTable::SetSpacingAt(int col, int row, float val) {
+void SpacingTable::SetSpacingAt(int col, int row, double val) {
     assert(col + n_col_ * row < n_col_ * n_row_);
     spacing_[col + n_col_ * row] = val;
 }
 
-float SpacingTable::GetParallelRunLengthAt(int col) const {
+double SpacingTable::GetParallelRunLengthAt(int col) const {
     assert(col < n_col_);
     return parallel_run_length_[col];
 }
 
-float SpacingTable::GetWidthAt(int row) const {
+double SpacingTable::GetWidthAt(int row) const {
     assert(row < n_row_);
     return width_[row];
 }
 
-float SpacingTable::GetSpacingAt(int col, int row) const {
+double SpacingTable::GetSpacingAt(int col, int row) const {
     assert(col + n_col_ * row < n_col_ * n_row_);
     return spacing_[col + n_col_ * row];
 }
